refactor gdnative-core to have a clearer module structure
Two new major modules exist now, containg previous code but in a different
"physical" location in the filesystem:
- core_types: all types used for interacting with Godot-types
- nativescript: all functions, types and traits used for creating and
  using the NativeScript extension.

The new module structure also allows to use `gdnative-core` without
nativescript support by disabling it as a default feature.
