+++ gdnative-core/Cargo.toml
@@ -10,18 +10,23 @@ workspace = ".."
 edition = "2018"
 
 [features]
+default = ["nativescript"]
 gd_test = ["approx"]
+nativescript = ["bitflags", "parking_lot"]
 
 [dependencies]
 gdnative-sys = { path = "../gdnative-sys", version = "0.8.1" }
 libc = "0.2"
-bitflags = "1.2"
+
 euclid = "0.20.1"
-parking_lot = "0.9.0"
+
 gdnative-impl-proc-macros = { path = "../impl/proc_macros", version = "=0.9.0" }
 
 # Used for float comparisons in godot tests
 approx = { version = "0.3.2", optional = true }
+
+bitflags = { version = "1.2", optional = true }
+parking_lot = { version = "0.9.0", optional = true }
 
 [build-dependencies]
 gdnative_bindings_generator = { path = "../bindings_generator", version = "0.8.1" }
