+++ gdnative-core/src/core_types/byte_array.rs
