+++ gdnative-core/src/core_types/error.rs
