+++ gdnative-core/src/core_types/mod.rs
@@ -1,0 +1,47 @@
+//! Types that represent core-datatypes of Godot.
+
+pub mod geom;
+
+pub mod access;
+pub mod byte_array;
+pub mod color;
+pub mod color_array;
+pub mod dictionary;
+pub mod error;
+pub mod float32_array;
+pub mod int32_array;
+pub mod node_path;
+pub mod point2;
+pub mod rid;
+pub mod string;
+pub mod string_array;
+pub mod typed_array;
+pub mod variant;
+pub mod variant_array;
+pub mod vector2;
+pub mod vector2_array;
+pub mod vector3;
+pub mod vector3_array;
+
+pub use geom::*;
+
+pub use access::*;
+pub use byte_array::*;
+pub use color::*;
+pub use color_array::*;
+pub use dictionary::Dictionary;
+pub use error::GodotError;
+pub use float32_array::*;
+pub use int32_array::*;
+pub use node_path::*;
+pub use point2::*;
+pub use rid::*;
+pub use string::*;
+pub use string_array::*;
+pub use typed_array::TypedArray;
+pub use variant::*;
+pub use variant_array::*;
+pub use vector2::*;
+pub use vector2_array::*;
+pub use vector3::*;
+pub use vector3_array::*;
