+++ gdnative-core/src/core_types/node_path.rs
