+++ gdnative-core/src/core_types/string.rs
