+++ gdnative-core/src/core_types/string_array.rs
