+++ gdnative-core/src/nativescript/init/property/accessor.rs
@@ -3,14 +3,12 @@
 use std::fmt::Debug;
 use std::marker::PhantomData;
 
+use crate::nativescript::user_data::UserData;
+use crate::nativescript::Map;
+use crate::nativescript::MapMut;
+use crate::nativescript::NativeClass;
 use crate::object::GodotObject;
-use crate::user_data::UserData;
-use crate::FromVariant;
-use crate::Map;
-use crate::MapMut;
-use crate::NativeClass;
-use crate::ToVariant;
-use crate::Variant;
+use crate::*;
 
 mod invalid;
