+++ gdnative-core/src/nativescript/mod.rs
@@ -1,0 +1,12 @@
+//! Types and functions related to the NativeScript extension of GDNative.
+
+mod macros;
+
+pub mod class;
+pub mod init;
+pub mod type_tag;
+pub mod user_data;
+
+pub use class::*;
+pub use init::*;
+pub use user_data::{Map, MapMut, UserData};
