+++ gdnative-core/src/nativescript/user_data.rs
@@ -69,7 +69,7 @@ use std::mem;
 use std::sync::Arc;
 use std::time::Duration;
 
-use crate::NativeClass;
+use crate::nativescript::NativeClass;
 
 /// Trait for customizable user-data wrappers.
 ///
